pyproject.toml
[tool.poetry]
name = "incident-bot"
version = "2.0.0"
description = "The open source incident management framework."
authors = ["Scott Hawkins <scott@echoboomer.net>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
alembic = "^1.13.2"
apscheduler = "^3.10.4"
atlassian-python-api = "^3.41.14"
emails = "^0.6"
fastapi = "^0.112.1"
opsgenie-sdk = "^2.1.5"
pdpyras = "^5.2.0"
psycopg2-binary = "^2.9.9"
pydantic = {extras = ["email"], version = "^2.8.2"}
pydantic-settings = {extras = ["yaml"], version = "^2.4.0"}
pyjwt = "^2.9.0"
pytest = "^8.3.2"
pytest-asyncio = "^0.23.8"
pytest-env = "^1.1.3"
pytest-mock = "^3.14.0"
pytest-sqlalchemy-mock = "^0.1.7"
python-dotenv = "^1.0.1"
python-json-logger = "^2.0.7"
requests = "^2.32.3"
requests-mock = "^1.12.1"
slack-bolt = "^1.20.0"
slack-sdk = "^3.31.0"
structlog = "^24.4.0"
sqlmodel = "^0.0.21"
uvicorn = "^0.30.6"
jinja2 = "^3.1.4"
python-multipart = "^0.0.9"
passlib = "^1.7.4"
bcrypt = "^4.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
